代理采集线程池调整.

+++ masker-proxypool-core/src/main/java/com/jiashunx/app/masker/proxypool/core/util/MDefaultThreadFactory.java
@@ -2,6 +2,7 @@ package com.jiashunx.app.masker.proxypool.core.util;
 
 import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 /**
  * 自定义线程工厂类.
@@ -10,9 +11,9 @@ import java.util.concurrent.atomic.AtomicInteger;
  */
 public class MDefaultThreadFactory implements ThreadFactory {
 
-    private static final AtomicInteger poolNumber = new AtomicInteger(1);
+    private static final AtomicLong poolNumber = new AtomicLong(1);
     private final ThreadGroup group;
-    private final AtomicInteger threadNumber = new AtomicInteger(1);
+    private final AtomicLong threadNumber = new AtomicLong(1);
     private final String namePrefix;
 
     public MDefaultThreadFactory() {
